feat: plug 상세 페이지 provider 연결 및 더미 데이터 교체

+++ lib/widgets/plug_power_info.dart
@@ -2,7 +2,7 @@ import 'package:cafe_plug_guardian_client/style.dart';
 import 'package:flutter/material.dart';
 
 class PlugPowerInfomattion extends StatelessWidget {
-  int assignPower, usedPower, realTimePower;
+  double assignPower, usedPower, realTimePower;
   String startTime, runningTime;
 
   PlugPowerInfomattion({
